pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "adversarial-resume-screening"
version = "2.0.0"
description = "Multi-step OpenEnv environment for adversarial resume screening with fraud detection."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Ishika"}
]
dependencies = [
    "openenv-core",
    "pydantic",
    "fastapi",
    "uvicorn",
    "openai",
    "python-dotenv"
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["server*", "models.py", "client.py", "data*"]
